Refactor: move base helpers and update imports
