+++ cogs/role_management/services/language_role_service.py
@@ -2,7 +2,12 @@
 
 from utils.database import database
 import logging
-from services.base import get_or_create_server_record, store_persistent_message, get_persistent_message, delete_persistent_message
+from utils.base import (
+    get_or_create_server_record,
+    store_persistent_message,
+    get_persistent_message,
+    delete_persistent_message,
+)
 from typing import Optional, Dict, List
 
 logger = logging.getLogger("RoleService")
